Add files via upload
The R file(combine_assembly v4.R) contains codes for cleaning and predicting the candidates for Election Pakistan 2018. The other files are data file used for predicting the candidates. I didn't win any position in the contest, however, it was nice experience with respect to predicting political/social/political targets